requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain_community
qdrant-client
python-dotenv
sqlalchemy
mysqlclient
pymysql
pypdf
sentence-transformers
streamlit
requests